refactoring, ints replaced by longs and bug fixing

+++ app/src/main/java/com/example/mytestapp/SieveC.kt
@@ -1,15 +1,15 @@
 package com.example.mytestapp
 
-external fun sieveC(n: Int): Int
+external fun sieveC(n: Long): Long
 
-external fun sieveResultsC(n: Int): ArrayList<Int>
+external fun sieveResultsC(n: Long): ArrayList<Long>
 
-external fun sieveParallelC(n: Int): Int
+external fun sieveParallelC(n: Long): Long
 
-external fun sieveResultsParallelC(n: Int): ArrayList<Int>
+external fun sieveResultsParallelC(n: Long): ArrayList<Long>
 
-external fun sieveEvenRemovedC(n: Int): Int
+external fun sieveEvenRemovedC(n: Long): Long
 
-external fun sieveEvenRemovedParallelC(n: Int): Int
+external fun sieveEvenRemovedParallelC(n: Long): Long
 
-external fun sieveBitArrayC(n: Int): Int
+external fun sieveBitArrayC(n: Long): Long
